requirements.txt
torch==1.11.0
transformers==4.20.1
tqdm
PyYAML
sentencepiece
protobuf==3.20.1
yargy==0.11.0
records